% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_pings.R
\name{graph_pings}
\alias{graph_pings}
\title{Graph pings}
\usage{
graph_pings(ping)
}
\arguments{
\item{ping}{a [data.frame()] from [process_pings_file()]}
}
\value{
a [ggplot2::ggplot()]
}
\description{
Graph pings
}
\examples{
#' data1 =
fs::path_package(
  package = "ping",
  "extdata/data_2023-11-20.txt") |>
process_pings_file()
graph_pings(data1)
}
