\name{PICS_IMPORT}
\alias{makeRangedDataOutput}
\alias{segReads}
\alias{segReadsList}
\alias{segReadsPE}
\alias{segReadsListPE}
\alias{setParaEM}
\alias{setParaPrior}

\title{
	PICS functions.
}

\description{
	Functions imported from PICS and exposed when loading PING.
	Refer to PICS documentation for further information.
}


\seealso{
	\link[PICS]{makeRangedDataOutput}
	\link[PICS]{segReads}
	\link[PICS]{segReadsList}
	\link[PICS]{setParaEM}
	\link[PICS]{setParaPrior}
	\link[PICS]{segReadsPE}
	\link[PICS]{segReadsListPE}
}
