% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ping-package.R
\docType{package}
\name{ping-package}
\alias{ping}
\alias{ping-package}
\title{ping: Graph ping data}
\description{
Graphs data collected from the `ping` command.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/d-morrison/ping}
  \item Report bugs at \url{https://github.com/d-morrison/ping/issues}
}

}
\keyword{internal}
