% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_pings_file.R
\name{process_pings_file}
\alias{process_pings_file}
\title{Import ping data}
\usage{
process_pings_file(file, start_time = NULL)
}
\arguments{
\item{file}{path to a file containing ping data}

\item{start_time}{optional start time to align the data}
}
\value{
a [tibble::tibble()]
}
\description{
Import ping data
}
\examples{
data1 =
fs::path_package(
  package = "ping",
  "extdata/data_2023-11-20.txt") |>
process_pings_file()
}
