% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.ping.R, R/graph_pings.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{session_info}
\alias{autoplot}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{ggplot2}{\code{\link[ggplot2]{autoplot}}}

  \item{sessioninfo}{\code{\link[sessioninfo]{session_info}}}
}}

