\name{summary}
\Rdversion{1.1}
\docType{methods}
\alias{summary-methods}
\alias{summary}
\alias{summary,ping-method}
\alias{summary,pingList-method}
\alias{summary,segReads-method}
\alias{summary,segReadsList-method}
\title{summary PING } 

\description{
This methods summary the objects of PING.}

\usage{
\S4method{summary}{ping}(object)
\S4method{summary}{pingList}(object)
\S4method{summary}{segReads}(object)
\S4method{summary}{segReadsList}(object)
}


\arguments{
\item{object}{Object returned from \code{\link{ping}} .}
}


\author{
Xuekui Zhang <\email{xzhang@stat.ubc.ca}
Sangsoon Woo, \email{swoo@fhcrc.org}
Raphael Gottardo <\email{raphael.gottardo@ircm.qc.ca}
}


\seealso{
\code{\link{show}}
}

\keyword{print}
