% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.ping.R
\name{autoplot.pings}
\alias{autoplot.pings}
\title{`autoplot()` method for `"pings"` objects}
\usage{
\method{autoplot}{pings}(object, log_y = FALSE, ...)
}
\arguments{
\item{object}{a `"pings"` object created using [process_pings_file()]}

\item{log_y}{whether to display the y axis with logarithmic spacing}

\item{...}{unused}
}
\value{
a [ggplot2::ggplot()]
}
\description{
`autoplot()` method for `"pings"` objects
}
\examples{
fs::path_package(
  package = "ping",
  "extdata/data_2023-11-20.txt") |>
process_pings_file() |>
autoplot()
}
