\name{PICS_IMPORT}
\alias{makeRangedDataOutput}
\alias{segReads}
\alias{segReadsList}

\title{
	PICS functions.
}

\description{
	Functions imported from PICS and exposed when loading PING.
	Refer to PICS documentation for further information.
}


\seealso{
	\link[PICS]{makeRangedDataOutput}
	\link[PICS]{segReads}
	\link[PICS]{segReadsList}
}
