\name{PICS_IMPORT}
\alias{makeRangedDataOutput}
\alias{segReads}
\alias{segReadsList}
\alias{segReadsPE}
\alias{segReadsListPE}
\alias{setParaEM}
\alias{setParaPrior}
\alias{bam2gr}

\title{
	PICS functions.
}

\description{
	Functions imported from PICS and exposed when loading PING.
	Refer to PICS documentation for further information.

\code{?PICS::segReads}

\code{?PICS::segReadsList}

\code{?PICS::segReadsPE}

\code{?PICS::segReadsListPE}

\code{?PICS::setParaEM}

\code{?PICS::setParaPrior}

\code{?PICS::bam2gr}
}


\seealso{
	\link[PICS]{makeRangedDataOutput}
	\link[PICS]{segReads}
	\link[PICS]{segReadsList}
	\link[PICS]{setParaEM}
	\link[PICS]{setParaPrior}
	\link[PICS]{segReadsPE}
	\link[PICS]{segReadsListPE}
}
