% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ping-package.R
\docType{package}
\name{ping-package}
\alias{ping}
\alias{ping-package}
\title{ping: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/d-morrison/ping}
  \item Report bugs at \url{https://github.com/d-morrison/ping/issues}
}

}
\keyword{internal}
